api/requirements.txt
fastapi
uvicorn
motor
dnspython
pyjwt[crypto]
python-dotenv
requests